Remove functions that should rather be implemented in objects.
All the carry transform operations should be handled by individual objects because the library cannot hope to provide any sensible defaults and should not do so.
In the worst of cases, objects would have to override these functions, just to get rid of them.

I also removed 'Collectible = 1' because that as well is something that should be kept in objects (and for erroneous creations of the library in a scenario, can turn out badly).