Added Ammo Logic Plugin. Implementing all of the ammo related functions previously in the library.
It should be unimportant in which order the plugin and the firearm library are included. However, the ammo manager should be included before the plugin.
With just the plugin, the weapon will take ammo from the configured ammo source object. No reloading mechanic is then present. This could be used to make Quake-style gameplay.
When using the 'reload from ammo source' plugin, however, the ammo manager should be included as well.