Format fire mode documentation
Removed the version tag in functions, because we deem it unnecessary.

Some infos for writing the docu, because that is not documented
anywhere:
The @br tag is necessary for formatting only, if you want to create a
line break.
The parameter type is read from the function signature, so the @par tag
should preferrable contain the meaning of the parameter.
The return value type cannot be decided easily be the parser (yet?) so
the @return tag allows an optional type description (same string as in
the parameter signature): "@return <descriptive text>" or "@return
<type> <descriptive text>"