Document some setters
Also added projectile spread again. Might remove "spread" in the future,
as it seems to be related to more specific behaviour: The "spread" for
continuous firing has to be calculated and added in the
GetSpread()-function in a custom implementation (which may be a plugin,
of course). This would also be the function where you apply spread from
movement, etc. by requesting spread values from the user.