Spread: Save progress
The original idea of having deviations as a variable value with a value
and precision was cool, because it provided a comfortable for adding a
specific deviation without caring about whether the values are
compatible to the rest.

On the script side this is still not too hard to handle (although I got
lazy with this commit), but the benefit from the additional calculations
seems very low. After all, in most implementations the user will pick
the same precision value anyway.