Reload: States progress
SOme progress, but this is still confused. On the one hand, the weapon
has to track its reloading state. On the other hand, the reload effect
has to track it (at least a little), but the most efficient way would be
to have the same effect for all stages instead of having one effect per
stage.